requirements.txt
langchain==0.3.13
langchain-community==0.3.13
pinecone-client==5.0.1
sentence-transformers==3.3.1
anthropic==0.39.0
python-dotenv==1.0.1
pypdf==5.1.0
numpy==2.2.0